Task in description
For the given interval [0, 1], weight function p(x)=sinx, n = 3 and number n,
construct a quadrature formula NAST with n + 1 nodes for
approximate calculation of the integral of the form ∫f(x)p(x)dx. Specify m - AST
of the constructed quadrature formula
2. Write out the formula for the remainder of the constructed CF:
3. Compare the accuracy of the constructed formula with formula B - a formula similar in
labor intensity of the Newton-Cotes type (at equally spaced
nodes, but with the weight function p(x)). The comparison is carried out for
the following types of integrand functions f:
a) polynomials of degree n, n+1, m, m+1 (m is the AST of the first constructed CF)
b) a smooth function (other than a polynomial)
c) a function with a discontinuity of the first derivative inside the segment
of integration.